# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CrisisMesh Operations API
  version: 0.1.0
  description: Normalized disaster incidents, response resources, and explainable allocations.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      operationId: health
      responses:
        "200": { description: Service is running }
  /api/v1/incidents:
    get:
      operationId: listIncidents
      parameters:
        - { name: kind, in: query, schema: { type: string } }
        - { name: severity, in: query, schema: { type: string, enum: [low, moderate, high, critical] } }
        - { name: status, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 } }
      responses:
        "200":
          description: Priority-ordered incidents
          content: { application/json: { schema: { $ref: "#/components/schemas/IncidentList" } } }
    post:
      operationId: ingestIncident
      security: [{ ApiKey: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/Incident" } } }
      responses:
        "201": { description: Incident accepted, content: { application/json: { schema: { $ref: "#/components/schemas/Incident" } } } }
        "422": { description: Invalid incident }
  /api/v1/incidents/{id}:
    get:
      operationId: getIncident
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Incident, content: { application/json: { schema: { $ref: "#/components/schemas/Incident" } } } }
        "404": { description: Unknown incident }
  /api/v1/resources:
    get:
      operationId: listResources
      responses:
        "200": { description: Resource inventory }
  /api/v1/allocations:
    get:
      operationId: listAllocations
      parameters: [{ name: incidentId, in: query, schema: { type: string } }]
      responses: { "200": { description: Allocation recommendations } }
    post:
      operationId: createAllocation
      security: [{ ApiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [incidentId]
              properties: { incidentId: { type: string }, resourceIds: { type: array, items: { type: string } } }
      responses: { "201": { description: Ranked response package }, "502": { description: Intelligence service unavailable } }
  /api/v1/summary:
    get:
      operationId: getSummary
      responses: { "200": { description: Operational rollup and source health } }
  /api/v1/stream:
    get:
      operationId: streamUpdates
      responses:
        "200": { description: Server-Sent Event stream, content: { text/event-stream: { schema: { type: string } } } }
components:
  securitySchemes:
    ApiKey: { type: apiKey, in: header, name: X-API-Key }
  schemas:
    Incident:
      type: object
      required: [title, latitude, longitude]
      properties:
        id: { type: string }
        externalId: { type: string }
        title: { type: string, minLength: 1 }
        kind: { type: string }
        severity: { type: string, enum: [low, moderate, high, critical] }
        status: { type: string }
        source: { type: string }
        sourceUrl: { type: string, format: uri }
        description: { type: string }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        startedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        riskScore: { type: number, minimum: 0, maximum: 100 }
        confidence: { type: number, minimum: 0, maximum: 1 }
        affectedPopulation: { type: integer, minimum: 0 }
        regions: { type: array, items: { type: string } }
    IncidentList:
      type: object
      required: [data, count]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Incident" } }
        count: { type: integer }